requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.16
django-import-export==3.0.1
et-xmlfile==1.1.0
gunicorn==20.1.0
MarkupPy==1.14
msoffcrypto-tool==5.0.0
numpy==1.21.6
odfpy==1.4.1
olefile==0.46
openpyxl==3.0.10
pandas==1.3.5
protobuf==4.21.7
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.5
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tablib==3.2.1
xlrd==2.0.1
xlwt==1.3.0
